pyproject.toml
[project]
name = "agentic-intelligence"
version = "0.1.0"
description = "Multi-agent system for market intelligence using AI agents"
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "playwright (>=1.53.0,<2.0.0)",
    "langchain-groq (>=0.3.6,<0.4.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
